chore(core): rename entity to graph
