+++ yogo/packages/core/src/modules/group/input.ts
@@ -1,5 +1,5 @@
 import { InputType, Field } from "type-graphql";
-import { NoteInput } from '../note/entity'
+import { NoteInput } from '../note/graph'
 
 @InputType()
 export class PushNoteInput implements Dawlet.IGroup.PushNoteInput {
